fix few warning on windows
git-svn-id: https://svn.resiprocate.org/rep/resiprocate@4701 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ main/sip/repro/AbstractDb.cxx
@@ -20,7 +20,7 @@ using namespace std;
 static void 
 encodeString( oDataStream& s, const Data& data )
 {
-   short len = data.size();
+   short len = (short)data.size();
    s.write( (char*)(&len) , sizeof( len ) );
    s.write( data.data(), len );
 }
